requirements.txt
asgiref==3.3.4
autopep8==1.5.7
colorama==0.4.4
dj-database-url==0.5.0
# Django==3.2.3
gunicorn==20.1.0
mysqlclient==2.0.3
Pillow==8.2.0
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
qrcode==6.1
segno==1.3.3
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
Unipath==1.1
whitenoise==5.2.0
